+++ workbench/libs/zunegfx/src/backends/cybergfx/rectangle/cybergfx_draw_rectangle_to_locked_board.c
@@ -62,22 +62,22 @@ void CybergfxDrawRectangleToLockedDrawingBoard(struct RenderContext *renderport,
                                        height, fill_pixel);
     } else {
       /* Pattern/gradient brushes require per-pixel sampling */
-      int row_width = (int)width;
+      WORD row_width = (WORD)width;
       ULONG *row_buffer =
-          row_width > 0 ? malloc((size_t)row_width * sizeof(ULONG)) : NULL;
+          row_width > 0 ? malloc((ULONG)row_width * sizeof(ULONG)) : NULL;
       if (!row_buffer) {
         return;
       }
 
       /* Process each row of the rectangle */
-      for (int row = 0; row < (int)height; ++row) {
+      for (WORD row = 0; row < (WORD)height; ++row) {
         WORD py = y + row;
-        int px = 0;
+        WORD px = 0;
 
         /* Process 4 pixels at a time using SIMD batch sampling */
         for (; px <= row_width - 4; px += 4) {
-          int px_coords[4];
-          for (int i = 0; i < 4; ++i) {
+          WORD px_coords[4];
+          for (WORD i = 0; i < 4; ++i) {
             px_coords[i] = x + px + i;
           }
 
@@ -88,14 +88,14 @@ void CybergfxDrawRectangleToLockedDrawingBoard(struct RenderContext *renderport,
           /* Write opaque pixels directly using vectorized packing (no alpha
            * blending) */
           UBYTE alpha[4] = {0xFF, 0xFF, 0xFF, 0xFF};
-          for (int i = 0; i < 4; ++i) {
+          for (WORD i = 0; i < 4; ++i) {
             row_buffer[px + i] = pack_argb32(alpha[i], fc_r[i], fc_g[i], fc_b[i]);
           }
         }
 
         /* Process remaining pixels one at a time */
         for (; px < row_width; ++px) {
-          int actual_x = x + px;
+          WORD actual_x = x + px;
           UBYTE fc_r, fc_g, fc_b, fc_a;
           SampleBrush(fill_brush, x, y, width, height, actual_x, py, &fc_r,
                       &fc_g, &fc_b, &fc_a);
@@ -106,7 +106,7 @@ void CybergfxDrawRectangleToLockedDrawingBoard(struct RenderContext *renderport,
 
         /* Write the entire row directly to the pixel buffer */
         ULONG *dest_row = pixels + (py * pitch_pixels) + x;
-        for (int i = 0; i < row_width; ++i) {
+        for (WORD i = 0; i < row_width; ++i) {
           dest_row[i] = row_buffer[i];
         }
       }
@@ -122,7 +122,7 @@ void CybergfxDrawRectangleToLockedDrawingBoard(struct RenderContext *renderport,
     /* Draw rectangle outline using direct pixel access */
     ULONG *pixels = (ULONG *)board->pixels;
     ULONG pitch_pixels = board->pitch / 4;
-    int line_width = (int)(border_width + 0.5f); /* Round to nearest int */
+    WORD line_width = (WORD)(border_width + 0.5f); /* Round to nearest int */
     if (line_width < 1)
       line_width = 1;
